133069 create insertMarkdownText as a hook
